user logout implemented
