+++ src/components/Modal/Modal.module.css
@@ -1,12 +1,78 @@
-.overlay {
+/* Lightbox */
+.lightbox {
+  z-index: 3;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   position: fixed;
   top: 0;
   left: 0;
   width: 100vw;
   height: 100vh;
+}
+
+.lightbox__overlay {
+  position: relative;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.lightbox__content {
+  position: absolute;
+  animation: emergeContent 170ms ease-out forwards;
+}
+
+.lightbox__button {
+  position: absolute;
   display: flex;
-  justify-content: center;
   align-items: center;
-  background-color: rgba(0, 0, 0, 0.25);
-  z-index: 1200;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  padding: 0;
+  margin: 0;
+  border: none;
+  border-radius: 50%;
+  background-color: transparent;
+  opacity: 1;
+  color: var(--title-black-color);
+  cursor: pointer;
+  transition: background-color var(--animation), color var(--animation), opacity var(--animation);
+  outline: none;
+
+  background-position: center;
+  background-size: 60%;
+  background-repeat: no-repeat;
+}
+
+.lightbox__buttonClose {
+  composes: lightbox__button;
+  top: 8px;
+  right: 8px;
+}
+
+.lightbox__buttonClose:hover,
+.lightbox__buttonClose:focus {
+  color: var(--btn-primary-color);
+  background-color: var(--btn-secondary-color);
+  opacity: 0.7;
+}
+
+.icon {
+  fill: currentColor;
+  display: inline-block;
+}
+
+@keyframes emergeContent {
+  0% {
+    transform: scale(0.7) translateY(70px);
+    opacity: 0.8;
+  }
+  100% {
+    transform: scale(1);
+    opacity: 1;
+  }
 }
