+++ src/components/Modal/index.js
@@ -1,1 +1,1 @@
-export { default } from './Modal.jsx';
+export { default } from './Modal';
