+++ src/components/ModalAddTransaction/ModalAddTransaction.module.css
@@ -1,7 +1,7 @@
 .modal {
   padding: 20px;
-  margin-top: 20px;
-  margin-bottom: 20px;
+  /* margin-top: 20px;
+  margin-bottom: 20px; */
   width: 540px;
   height: 508px;
   border-radius: 20px;
