+++ src/redux/mainAPISlice.js
@@ -2,7 +2,7 @@ import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
 export const emptySplitApi = createApi({
   baseQuery: fetchBaseQuery({
-    baseUrl: 'https://wallet-proj.herokuapp.com',
+    baseUrl: 'https://wallet-proj.herokuapp.com/api',
     prepareHeaders: (headers, { getState }) => {
       const token = getState().auth?.token;
